error handling + code refactor

+++ Book_Catalogue/src/main/java/pl/rabowski/book_catalogue/service/BookService.java
@@ -10,7 +10,7 @@ public interface BookService {
 
 	public void setList(List<Book> list);
 
-	public Book getBook(int bookIndex);
+	public Book getBook(long bookIndex);
 
 	public void addBook(Book book);
